backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.5.0
python-dotenv==1.0.1
openai==1.3.0
python-multipart==0.0.6
korean-lunar-calendar==0.3.1
httpx==0.25.0
requests

# ORM + Supabase (Phase 0)
sqlmodel[asyncio]>=0.0.14
asyncpg>=0.29.0
greenlet>=3.0.0
psycopg2-binary>=2.9.9
